+++ src/Entity/Task.php
@@ -40,6 +40,16 @@ class Task
      */
     private $isDone;
 
+    /**
+     * @ORM\ManyToOne(targetEntity=User::class, inversedBy="tasks")
+     */
+    private $user;
+
+    /**
+     * @ORM\Column(type="datetime", nullable=true)
+     */
+    private $last_modification;
+
     public function __construct()
     {
         $this->createdAt = new \Datetime();
@@ -94,5 +104,29 @@ class Task
     public function setId(int $id)
     {
         $this->id = $id;
+    }
+
+    public function getUser(): ?User
+    {
+        return $this->user;
+    }
+
+    public function setUser(?User $user): self
+    {
+        $this->user = $user;
+
+        return $this;
+    }
+
+    public function getLastModification(): ?\DateTimeInterface
+    {
+        return $this->last_modification;
+    }
+
+    public function setLastModification(?\DateTimeInterface $last_modification): self
+    {
+        $this->last_modification = $last_modification;
+
+        return $this;
     }
 }