+++ tests/User/UserTest.php
@@ -31,4 +31,36 @@ class UserTest extends WebTestCase
         $this->assertSame($email, $user->getEmail());
         $this->assertSame($roles, $user->getRoles());
     }
+
+    public function testLogin()
+    {
+        $client = static::createClient();
+        //on arrive sur la page d'accueil
+        $crawler = $client->request('GET', '/');
+
+        //on clique sur se connecter
+        $link = $crawler->selectLink('Se connecter')->link();
+        $crawler = $client->click($link);
+
+        // ce qui doit être visible
+        $this->assertSame(1, $crawler->filter('html:contains("Nom d\'utilisateur")')->count());
+        $this->assertSame(1, $crawler->filter('html:contains("Mot de passe")')->count());
+
+        //quand on clique sur
+        $form = $crawler->selectButton('Se connecter')->form();
+        $form['username'] = 'username1';
+        $form['password'] = 'password1';
+        // submit
+        $client->submit($form);
+
+        //on affiche le contenu de la reponse
+        //on reprend la variable crawler pour récupéré le dernier script
+        $crawler = $client->followRedirect();
+
+        // on affiche le script
+        //echo $client->getResponse()->getContent();
+
+        //on verifie le dernier script
+        $this->assertSame(1, $crawler->filter('html:contains("Vous êtes connecté en tant que username1")')->count());
+    }
 }